requirements.txt
aiodns==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alabaster==0.7.16
anyio==4.12.1
appdirs==1.4.4
argcomplete==3.6.3
arrow==1.3.0
astroid==3.3.8
asttokens==3.0.0
asyncssh==2.21.0
attrs==25.4.0
autopep8==2.3.2
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.14.3
binaryornot==0.4.4
black==25.1.0
bleach==6.2.0
blivet==3.12.1
blivet-gui==2.6.0
blosc2==3.8.0
boto3==1.42.40
botocore==1.42.40
Bottleneck==1.6.0
Brlapi==0.8.7
Brotli==1.2.0
cachetools==5.5.2
cattrs==25.3.0
certifi==2025.7.9
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.7
cloudpickle==3.1.2
cockpit @ file:///builddir/build/BUILD/cockpit-354-build/cockpit-354/tmp/wheel/cockpit-354-py3-none-any.whl#sha256=1d4fb1517c54d5d11d999be3583ea5fbf3d3d011cb789fb717daae0672c38816
comm==0.2.3
contourpy==1.3.3
cookiecutter==2.6.0
crypt_r==3.13.1
cryptography==45.0.4
cssselect==1.3.0
cupshelpers==1.0
cycler==0.11.0
dasbus==1.7
dbus-python==1.4.0
dbus_next==0.2.3
decorator==5.2.1
defusedxml==0.7.1
diff-match-patch==20241021
dill==0.4.1
distro==1.9.0
dnf==4.24.0
docstring-to-markdown==0.17
docutils==0.21.2
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
fedora-third-party==0.10
file-magic==0.4.0
flake8==6.1.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2026.1.0
gcsfs==2025.9.0
google-api-core==2.27.0
google-auth==2.45.0
google-auth-oauthlib==1.2.3
google-cloud-core==2.3.3
google-cloud-storage==2.14.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.63.0
greenlet==3.1.1
grpcio==1.48.4
grpcio-status==1.48.4
h11==0.16.0
helpdev==0.7.1
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
idna==3.10
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
imagesize==1.4.1
importlib_metadata==8.7.0
inflection==0.5.1
initial-setup==0.3.101
inkex==1.4.0
intervaltree==3.1.0
ipykernel==6.29.3
ipython==8.37.0
ipython_pygments_lexers==1.1.1
iso639==0.1.4
isort==6.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.9.0
jellyfish==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.1
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
keyring==25.7.0
kiwisolver==1.4.9
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.17.0
libsass==0.23.0
libvirt-python==11.6.0
louis==3.33.0
lsprotocol==2025.0.0
lxml==6.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.5.0
moviepy==2.2.1
msgpack==1.1.0
multidict==6.6.4
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
ndindex==1.10.0
nest_asyncio==1.6.0
netifaces==0.11.0
nftables==0.1
nodejs-wheel-binaries==24.13.0
numexpr==2.12.1
numpy==2.3.5
numpydoc==1.8.0
oauthlib==3.2.2
odfpy==1.4.1
olefile==0.47
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==1.0.3
pexpect==4.9.0
pickleshare==0.7.5
pid==3.0.4
pillow==11.3.0
platformdirs==4.2.2
pluggy==1.6.0
ply==3.11
pooch==1.8.2
productmd==1.50
proglog==0.1.12
prompt_toolkit==3.0.41
propcache==0.4.1
proto-plus==1.22.3
protobuf==3.19.6
psutil==7.0.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.13
pycairo==1.28.0
pycares==4.10.0
pycodestyle==2.12.1
pycparser==2.22
pycups==2.0.4
pydocstyle==6.3.0
pyenchant==3.2.2
pyflakes==3.1.0
pygame==2.6.1
pygdbmi==0.11.0.0
PyGithub==2.7.0
Pygments==2.19.1
PyGObject==3.54.5
PyJWT==2.8.0
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
pylint==3.3.9
pylint-venv==3.0.4
pyls-spyder==0.4.0
PyMySQL==1.1.2
PyNaCl==1.5.0
pyparsing==3.1.2
pyparted==3.13.0
PyQt5==5.15.12.dev2507081429
PyQt5_sip==12.17.1
PyQtWebEngine==5.15.6
pyserial==3.5
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-ruff==2.3.0
python-lsp-server==1.13.1
python-meh==0.52
python-slugify==8.0.4
pytoolconfig==1.3.1
pytubefix==10.3.6
pytz==2025.2
pyuca==1.2
pyudev==0.24.3
pyxdg==0.28
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.18.8-build/kernel-6.18.8/linux-6.18.8-200.fc43.x86_64/tools/net/ynl
pyzmq==27.1.0
QDarkStyle==3.2.3
qstylizer==0.2.4
QtAwesome==1.4.0
qtconsole==5.7.0
QtPy==2.4.3
qtsass==0.4.0
RapidFuzz==3.12.2
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
requests-oauthlib==1.3.1
rich==14.1.0
roman-numerals-py==3.1.0
rope==1.14.0
rpds-py==0.27.0
rpm==6.0.1
rsa==4.9
rtree==1.4.1
ruff==0.14.13
s3transfer==0.16.0
scipy==1.16.2
scour==0.38.2
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
sentry-sdk==2.35.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.36-build/setroubleshoot-3.3.36/src
setuptools==78.1.1
shtab==1.7.2
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
sos==4.10.2
soupsieve==2.8
Sphinx==8.2.3
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder==6.1.0
spyder-kernels==3.1.0
SQLAlchemy==2.0.45
stack_data==0.6.3
superqt==0.7.7
systemd-python==235
tables==3.10.2
tabulate==0.9.0
text-unidecode==1.3
textdistance==4.6.3
three-merge==0.1.1
tinycss2==1.5.1
tomli==2.2.1
tomlkit==0.13.2
tornado==6.5.2
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
whatthepatch==1.0.7
wurlitzer==3.1.1
xarray==2025.12.0
xkbregistry==0.3
xlrd==2.0.1
xlsxwriter==3.2.9
yapf==0.40.2
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
